Bug Fixes
Bug where old save handlers were not being unbound when the model popup
was closed. This means that any items that were viewed and closed in the
dashboard would be updated if the user updates an item later. Fixed with
using jQuery.off('click') to unbind any residual handlers before each
jQuery.on('click') event binding.